+++ ProiectSemestru/src/main/java/project/moto/Repository/DatasbaseRepository/PlayerDBRepository.java
@@ -1,6 +1,10 @@
-package project.moto.Repository;
+package project.moto.Repository.DatasbaseRepository;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import project.moto.Domain.Player;
+import project.moto.Repository.PlayerRepository;
+import project.moto.Utils.JdbcUtils;
 
 
 import java.sql.Connection;
@@ -8,32 +12,27 @@ import java.sql.PreparedStatement;
 import java.util.*;
 
 
-public class PlayerDBRepository extends AbstractDatabaseRepository<Integer,Player>
+public class PlayerDBRepository implements PlayerRepository
 {
-
+    private final JdbcUtils dbUtils;
+    private static final Logger logger = LogManager.getLogger();
     public PlayerDBRepository(Properties props) {
-        super(props);
-        Load();
-        logger.info("Initialized PlayerDBIRepository with properties: {}", props);
+        dbUtils = new JdbcUtils(props);
+        logger.info("Initialized PlayerDBRepository with properties: {}", props);
     }
 
     /**
-     * Find the entity with the given id, if it exists in memory in returns instantly, else queries ithe DBs
-     * @param aInt -the id of the entity to be returned
-     * id must not be null
+     * Find the entity with the given id
+     * @param pId -the id of the entity to be returned
+     * pId must not be null
      * @return Player - null if it doesnt exist
      */
     @Override
-    public Optional<Player> findOne(Integer aInt) {
-        logger.info("Finding Player by ID: " + aInt);
-        if(data.containsKey(aInt))
-        {
-            logger.traceExit("Player MEMORY_HIT");
-            return Optional.of(data.get(aInt));
-        }
+    public Optional<Player> findOne(Integer pId) {
+        logger.info("Finding Player by ID: " + pId);
         Connection con = dbUtils.getConnection();
         try (var preStmt = con.prepareStatement("select * from Player where id = ?")) {
-            preStmt.setInt(1, aInt);
+            preStmt.setInt(1, pId);
             try (var result = preStmt.executeQuery()) {
                 if (result.next()) {
                     int id = result.getInt("Id");
@@ -42,8 +41,7 @@ public class PlayerDBRepository extends AbstractDatabaseRepository<Integer,Playe
                     Integer team = result.getInt("Team");
                     Player player = new Player(name, code, team);
                     player.setId(id);
-                    logger.traceExit();
-                    data.put(id, player);
+                    logger.traceExit("Player found");
                     return Optional.of(player);
                 }
             }
@@ -51,22 +49,19 @@ public class PlayerDBRepository extends AbstractDatabaseRepository<Integer,Playe
             logger.error(e.getMessage());
             throw new RuntimeException(e);
         }
-        logger.traceExit();
+        logger.traceExit("Player not found");
         return Optional.empty();
     }
 
+    /**
+     * Find all entities in the database
+     * @return Map<Integer,Player> - all entities
+     */
     @Override
     public Map<Integer,Player> findAll() {
         logger.info("Finding all Players");
-        return super.findAll();
-    }
-
-    /**
-     * Load all Players from the database
-     */
-    @Override
-    protected void Load(){
         Connection con = dbUtils.getConnection();
+        Map<Integer,Player> data = new HashMap<>();
         try (var preStmt = con.prepareStatement("select * from Player")) {
             try (var result = preStmt.executeQuery()) {
                 while (result.next()) {
@@ -78,16 +73,19 @@ public class PlayerDBRepository extends AbstractDatabaseRepository<Integer,Playe
                     player.setId(id);
                     data.put(id,player);
                 }
+
             }
         } catch (Exception e) {
             logger.error(e.getMessage());
             throw new RuntimeException(e);
         }
+
         logger.traceExit("Found {} Players in database", data.size());
+        return data;
     }
 
     /**
-     * Save the entity to the database and locally
+     * Save the entity to the database
      * @param entity
      * entity must be not null
      * @return entity if the entity was saved, null otherwise
@@ -109,7 +107,6 @@ public class PlayerDBRepository extends AbstractDatabaseRepository<Integer,Playe
                         entity.setId(generatedKeys.getInt(1));
                     }
                 }
-                data.put(entity.getId(), entity);
                 logger.traceExit("Player saved with id" + entity.getId());
                 return Optional.of(entity);
             }
@@ -123,28 +120,33 @@ public class PlayerDBRepository extends AbstractDatabaseRepository<Integer,Playe
     }
 
     /**
-     * Delete the entity from the database and locally
-     * @param aInt
+     * Delete the entity from the database
+     * @param pID
      * id must be not null
      * @return the deleted entity if the entity was deleted, null otherwise
      */
     @Override
-    public Optional<Player> delete(Integer aInt) {
-        logger.info("Deleting Player: {}", aInt);
+    public Optional<Player> delete(Integer pID) {
+        logger.info("Deleting Player: {}", pID);
         Connection con = dbUtils.getConnection();
-        try (var preStmt = con.prepareStatement("delete from Player where id = ?")) {
-            preStmt.setInt(1, aInt);
-            preStmt.executeUpdate();
-            logger.traceExit();
-            return Optional.of(data.remove(aInt));
-        } catch (Exception e) {
-            logger.error(e.getMessage());
-            throw new RuntimeException(e);
+        Optional<Player> player = findOne(pID);
+        if(player.isPresent()) {
+            try (var preStmt = con.prepareStatement("delete from Player where id = ?")) {
+                preStmt.setInt(1, pID);
+                preStmt.executeUpdate();
+                logger.traceExit("Player {} deleted", pID);
+                return player;
+            } catch (Exception e) {
+                logger.error(e.getMessage());
+                throw new RuntimeException(e);
+            }
         }
+        logger.traceExit("Player was not deleted");
+        return Optional.empty();
     }
 
     /**
-     * Update the entity in the database and locally
+     * Update the entity in the database
      * @param entity - the entity to update
      * @return old entity if the entity was updated, null otherwise
      */
@@ -158,8 +160,8 @@ public class PlayerDBRepository extends AbstractDatabaseRepository<Integer,Playe
             preStmt.setInt(3, entity.getTeam());
             preStmt.setInt(4, entity.getId());
             preStmt.executeUpdate();
-            logger.traceExit();
-            return Optional.ofNullable(data.put(entity.getId(), entity));
+            logger.traceExit("Player updated succesfully");
+            return Optional.of(entity);
         } catch (Exception e) {
             logger.error(e.getMessage());
             throw new RuntimeException(e);
