+++ ProiectSemestru/src/main/java/project/moto/Repository/DatasbaseRepository/TeamDBRepository.java
@@ -1,45 +1,42 @@
-package project.moto.Repository;
+package project.moto.Repository.DatasbaseRepository;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import project.moto.Domain.Team;
+import project.moto.Repository.TeamRepository;
+import project.moto.Utils.JdbcUtils;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.util.*;
 
-public class TeamDBRepository extends AbstractDatabaseRepository<Integer, Team> {
-
+public class TeamDBRepository implements TeamRepository {
+    private final JdbcUtils dbUtils;
+    private static final Logger logger = LogManager.getLogger();
     public TeamDBRepository(Properties props) {
-        super(props);
-        Load();
+        this.dbUtils = new JdbcUtils(props);
         logger.info("Initialized TeamDBRepository with properties: {}", props);
 
     }
 
     /**
-     * Find the Team with the given id, if it exists in memory in returns instantly, else queries the DB
-     * @param aInt -the id of the team to be returned
+     * Find the Team with the given id
+     * @param tID -the id of the team to be returned
      * id must not be null
      * @return Team - null if it doesnt exist
      */
     @Override
-    public Optional<Team> findOne(Integer aInt) {
-        logger.info("Finding Team by ID: {}", aInt);
-        if(data.containsKey(aInt))
-        {
-            logger.traceExit("Team MEMORY_HIT");
-            return Optional.of(data.get(aInt));
-        }
+    public Optional<Team> findOne(Integer tID) {
         Connection con = dbUtils.getConnection();
         try (var preStmt = con.prepareStatement("select * from Team where id = ?")) {
-            preStmt.setInt(1, aInt);
+            preStmt.setInt(1, tID);
             try (var result = preStmt.executeQuery()) {
                 if (result.next()) {
                     int id = result.getInt("Id");
                     String name = result.getString("Name");
                     Team team = new Team(name);
                     team.setId(id);
-                    logger.traceExit();
-                    data.put(id, team);
+                    logger.traceExit("Team Found");
                     return Optional.of(team);
                 }
             }
@@ -57,17 +54,9 @@ public class TeamDBRepository extends AbstractDatabaseRepository<Integer, Team> 
      */
     @Override
     public Map<Integer,Team> findAll() {
-        logger.info("Finding all Teams");
-        return super.findAll();
-    }
-
-    /**
-     * Load all Teams from the database
-     */
-    @Override
-    protected void Load(){
-        logger.info("Loading Teams from database");
+        logger.info("Finding Teams from database");
         Connection con = dbUtils.getConnection();
+        Map<Integer,Team> data = new HashMap<>();
         try (var preStmt = con.prepareStatement("select * from Team")) {
             try (var result = preStmt.executeQuery()) {
                 while (result.next()) {
@@ -77,16 +66,17 @@ public class TeamDBRepository extends AbstractDatabaseRepository<Integer, Team> 
                     team.setId(id);
                     data.put(id,team);
                 }
+                logger.traceExit("Found {} Teams",data.size());
+                return data;
             }
         } catch (Exception e) {
             logger.error(e.getMessage());
             throw new RuntimeException(e);
         }
-        logger.traceExit("Found {} Teams",data.size());
     }
 
     /**
-     * Save a Team to the database and locally
+     * Save a Team to the database
      * @param entity
      * entity must be not null
      * @return Optional<Team> - null if the Team was not successfully saved
@@ -105,7 +95,6 @@ public class TeamDBRepository extends AbstractDatabaseRepository<Integer, Team> 
                         entity.setId(generatedKeys.getInt(1));
                     }
                 }
-                data.put(entity.getId(), entity);
                 logger.traceExit("Team {} saved successfully", entity.getId());
                 return Optional.of(entity);
             }
@@ -118,23 +107,29 @@ public class TeamDBRepository extends AbstractDatabaseRepository<Integer, Team> 
     }
 
     /**
-     * Delete a Team from the database and locally
-     * @param integer
+     * Delete a Team from the database
+     * @param id
      * id must be not null
      * @return Optional<Team> - null if the Team was not successfully deleted
      */
     @Override
-    public Optional<Team> delete(Integer integer) {
-        logger.traceEntry("deleting team {} ", integer);
+    public Optional<Team> delete(Integer id) {
+        logger.traceEntry("deleting team {} ", id);
         Connection con = dbUtils.getConnection();
-        try (var preStmt = con.prepareStatement("delete from Team where id = ?")) {
-            preStmt.setInt(1, integer);
-            preStmt.executeUpdate();
-            logger.traceExit("Deleted team {}", integer);
-            return Optional.of(data.remove(integer));
-        } catch (Exception e) {
-            logger.error(e.getMessage());
-            throw new RuntimeException(e);
+        Optional<Team> team = findOne(id);
+        if(team.isPresent()) {
+            try (var preStmt = con.prepareStatement("delete from Team where id = ?")) {
+                preStmt.setInt(1, id);
+                preStmt.executeUpdate();
+                logger.traceExit("Deleted team {}", id);
+                return team;
+            } catch (Exception e) {
+                logger.error(e.getMessage());
+                throw new RuntimeException(e);
+            }
+        }else{
+            logger.traceExit("Team not found");
+            return Optional.empty();
         }
     }
 
@@ -152,7 +147,7 @@ public class TeamDBRepository extends AbstractDatabaseRepository<Integer, Team> 
             preStmt.setInt(2, entity.getId());
             preStmt.executeUpdate();
             logger.traceExit("Successfully updated team {} data", entity.getId());
-            return Optional.ofNullable(data.put(entity.getId(), entity));
+            return Optional.of(entity);
 
         } catch (Exception e) {
             logger.error(e.getMessage());
